working write up except for docker image
